Fixed bug on taking transpose of dual field over refined mesh
The old method for taking the transpose of a dual field just used CG to
compute the inverse of the mass matrix times the dual field. This code
wasn't condensing or distributing hanging node constraints, so we would
get wrong results when applying it to dual fields over meshes with
hanging nodes.