Changed ownership policies for discretization and triangulation
This patch makes `Discretization` only usable through a
`std::shared_ptr`. Before, we could create discretization objects
somewhere on the stack by invoking the class's constructor; fields and
vector fields would refer to a discretization through a
`dealii::SmartPointer`, which would alert the user if the discretization
was deleted prematurely. This works perfectly fine in C++ because of
RAII, but as soon as we start binding icepack to languages like python
with a different memory model, all the assumptions that make this
strategy work are no longer true. By making fields and vector fields
share ownership of the discretization, we guarantee memory consistency
at negligible performance cost.

Likewise, discretization objects now own the underlying triangulation.
The mesh is copied upon construction of the discretization; this is
inefficient and I should probably overload it for move semantics.